+++ src/gui/gui-buffer.c
@@ -85,7 +85,7 @@ char *gui_buffer_properties_get_integer[] =
 { "opening", "number", "old_number", "layout_number", "layout_number_merge_order",
   "type", "notify", "num_displayed", "active", "hidden", "zoomed",
   "print_hooks_enabled", "day_change", "clear", "filter", "closing",
-  "lines_hidden", "prefix_max_length", "next_line_id", "time_for_each_line",
+  "lines_hidden", "prefix_max_length", "time_for_each_line",
   "nicklist", "nicklist_case_sensitive", "nicklist_max_length",
   "nicklist_display_groups", "nicklist_count", "nicklist_visible_count",
   "nicklist_groups_count", "nicklist_groups_visible_count",
@@ -98,7 +98,7 @@ char *gui_buffer_properties_get_integer[] =
   NULL
 };
 char *gui_buffer_properties_get_longlong[] =
-{ "id", "nicklist_last_id_assigned",
+{ "id", "lines_last_id_assigned", "nicklist_last_id_assigned",
   NULL
 };
 char *gui_buffer_properties_get_string[] =
@@ -891,7 +891,7 @@ gui_buffer_new_props_with_id (long long id,
     new_buffer->own_lines = gui_line_lines_alloc ();
     new_buffer->mixed_lines = NULL;
     new_buffer->lines = new_buffer->own_lines;
-    new_buffer->next_line_id = 0;
+    new_buffer->lines_last_id_assigned = -1;
     new_buffer->time_for_each_line = 1;
     new_buffer->chat_refresh_needed = 2;
 
@@ -1462,8 +1462,6 @@ gui_buffer_get_integer (struct t_gui_buffer *buffer, const char *property)
         return buffer->lines->lines_hidden;
     else if (strcmp (property, "prefix_max_length") == 0)
         return buffer->lines->prefix_max_length;
-    else if (strcmp (property, "next_line_id") == 0)
-        return buffer->next_line_id;
     else if (strcmp (property, "time_for_each_line") == 0)
         return buffer->time_for_each_line;
     else if (strcmp (property, "nicklist") == 0)
@@ -1536,6 +1534,8 @@ gui_buffer_get_longlong (struct t_gui_buffer *buffer, const char *property)
 
     if (strcmp (property, "id") == 0)
         return buffer->id;
+    else if (strcmp (property, "lines_last_id_assigned") == 0)
+        return buffer->lines_last_id_assigned;
     else if (strcmp (property, "nicklist_last_id_assigned") == 0)
         return buffer->nicklist_last_id_assigned;
 
@@ -5386,7 +5386,7 @@ gui_buffer_hdata_buffer_cb (const void *pointer, void *data,
         HDATA_VAR(struct t_gui_buffer, own_lines, POINTER, 0, NULL, "lines");
         HDATA_VAR(struct t_gui_buffer, mixed_lines, POINTER, 0, NULL, "lines");
         HDATA_VAR(struct t_gui_buffer, lines, POINTER, 0, NULL, "lines");
-        HDATA_VAR(struct t_gui_buffer, next_line_id, INTEGER, 0, NULL, NULL);
+        HDATA_VAR(struct t_gui_buffer, lines_last_id_assigned, LONGLONG, 0, NULL, NULL);
         HDATA_VAR(struct t_gui_buffer, time_for_each_line, INTEGER, 0, NULL, NULL);
         HDATA_VAR(struct t_gui_buffer, chat_refresh_needed, INTEGER, 0, NULL, NULL);
         HDATA_VAR(struct t_gui_buffer, nicklist, INTEGER, 0, NULL, NULL);
@@ -5600,7 +5600,7 @@ gui_buffer_add_to_infolist (struct t_infolist *infolist,
         return 0;
     if (!infolist_new_var_integer (ptr_item, "prefix_max_length", buffer->lines->prefix_max_length))
         return 0;
-    if (!infolist_new_var_integer (ptr_item, "next_line_id", buffer->next_line_id))
+    if (!infolist_new_var_longlong (ptr_item, "lines_last_id_assigned", buffer->lines_last_id_assigned))
         return 0;
     if (!infolist_new_var_integer (ptr_item, "time_for_each_line", buffer->time_for_each_line))
         return 0;
@@ -5849,7 +5849,7 @@ gui_buffer_print_log (void)
         log_printf ("  mixed_lines . . . . . . : %p", ptr_buffer->mixed_lines);
         gui_lines_print_log (ptr_buffer->mixed_lines);
         log_printf ("  lines . . . . . . . . . : %p", ptr_buffer->lines);
-        log_printf ("  next_line_id. . . . . . : %d", ptr_buffer->next_line_id);
+        log_printf ("  lines_last_id_assigned. : %lld", ptr_buffer->lines_last_id_assigned);
         log_printf ("  time_for_each_line. . . : %d", ptr_buffer->time_for_each_line);
         log_printf ("  chat_refresh_needed . . : %d", ptr_buffer->chat_refresh_needed);
         log_printf ("  nicklist. . . . . . . . : %d", ptr_buffer->nicklist);